use PATH_CACHE consistently within determinator.rb
Previously, `PATH_CACHE` was only used by `formatted_last_modified_date` and
not by `last_modified_at`. This resulted in last modified time being
re-determined when using:

{{ page.last_modified_at }}

but not when using:

{% last_modified_at %}

This change aligns the behavior and, as a by product, improves site reset
render performance for users of `page.last_modified_at`. Initial site
generation render performance is unaffected. The tradeoff is that data can get
stale in the cache, and repo information is not reread on reset.